requirements.txt
Flask==2.0.2
PyNaCl==1.4.0
requests==2.26.0
